+++ src/cli/commands/run.js
@@ -6,6 +6,7 @@ import {execCommand, makeEnv} from '../../util/execute-lifecycle-script.js';
 import {dynamicRequire} from '../../util/dynamic-require.js';
 import {MessageError} from '../../errors.js';
 import {registries} from '../../resolvers/index.js';
+import {checkOne as checkCompatibility} from '../../package-compatibility.js';
 import * as fs from '../../util/fs.js';
 import * as constants from '../../constants.js';
 
@@ -118,6 +119,13 @@ export async function run(config: Config, reporter: Reporter, flags: Object, arg
     }
 
     if (cmds.length) {
+      const ignoreEngines = !!(flags.ignoreEngines || config.getOption('ignore-engines'));
+      try {
+        await checkCompatibility(pkg, config, ignoreEngines);
+      } catch (err) {
+        throw err instanceof MessageError ? new MessageError(reporter.lang('cannotRunWithIncompatibleEnv')) : err;
+      }
+
       // Disable wrapper in executed commands
       process.env.YARN_WRAP_OUTPUT = 'false';
       for (const [stage, cmd] of cmds) {